Refactor slightly
Split types and operations representing JSON Pointer and JSON Patch
into separate modules.

Drop the value from the remove operation constructor.
