+++ aeson-diff.cabal
@@ -1,5 +1,5 @@
 name:                aeson-diff
-version:             1.0.0.0
+version:             1.0.0.1
 synopsis:            Extract and apply patches to JSON documents.
 description:
   .
@@ -32,6 +32,8 @@ library
   default-language:    Haskell2010
   hs-source-dirs:      lib
   exposed-modules:     Data.Aeson.Diff
+                     , Data.Aeson.Patch
+                     , Data.Aeson.Pointer
   build-depends:       base >=4.5 && <4.9
                      , aeson
                      , bytestring >= 0.10
